extract task portability work to separate file
